GLScene1, GLScene2 , WebGPUScene1, WebGPUScene2

+++ src/index.ts
@@ -1,39 +1,45 @@
 import "./styles/styles.scss";
 
 import { GUI } from "lil-gui";
-import { canvasHeight, canvasWidth, HideNoSupport } from "src/general";
 import Stats from "stats.js";
 
-import { GlScene1 } from "./gl/Scene1GL";
-import { timer } from "./profiler";
-import { Scene } from "./shared/scene";
+import { RenderedSceneInfo } from "./interfaces/RenderedSceneInfo";
+import { Scene } from "./interfaces/Scene";
 import { buildCanvas, scrollFadeOut } from "./utilities/CanvasBuilder";
-import { BuildScene, RendererScenesStrings } from "./utilities/SceneBuilder";
+import { Profiler } from "./utilities/Profiler";
+import { BuildScene, initRenderer, initSceneName, RendererScenesStrings } from "./utilities/SceneBuilder";
+import Timer from "./utilities/Timer";
 
 const gui = new GUI();
 let canvas = buildCanvas();
-canvas.width = canvasWidth;
-canvas.height = canvasHeight;
-HideNoSupport();
 
-let scene: Scene = new GlScene1();
-
-const RenderedScene: { sceneName: string; renderer: "webgl2" | "webgpu" } = {
-    sceneName: "scene1",
-    renderer: "webgl2"
+let scene: Scene | undefined = undefined;
+const RenderedScene: RenderedSceneInfo = {
+    sceneName: initSceneName,
+    renderer: initRenderer
 };
-const rendering = gui.addFolder("Rendering");
+let sceneInitialized = false;
 
-const changeScene = () => {
+const changeScene = async () => {
+    sceneInitialized = false;
+    Profiler.sceneName = RenderedScene.sceneName;
+    Profiler.renderer = RenderedScene.renderer;
+
     canvas.remove();
     canvas = buildCanvas();
-    HideNoSupport();
-    scene.delete();
+    if (scene) {
+        scene.delete();
+    }
     scene = BuildScene(RenderedScene.renderer, RenderedScene.sceneName);
-    scene.init(canvas, gui);
+    scene.init(canvas, gui).then(() => {
+        sceneInitialized = true;
+    });
 };
+
+const rendering = gui.addFolder("Rendering");
 rendering
     .add(RenderedScene, "renderer", ["webgl2", "webgpu"])
+    .name("Renderer")
     .onFinishChange(() => {
         changeScene();
         guiScenes.options(RendererScenesStrings(RenderedScene.renderer));
@@ -45,35 +51,33 @@ const guiScenes = rendering
         "sceneName",
         RendererScenesStrings(RenderedScene.renderer)
     )
+    .name("Scene Name")
     .onFinishChange(() => {
         changeScene();
     });
 
 const profiler = gui.addFolder("Profiler");
+profiler.add(Profiler, "profilerExecutionInSeconds", 1, 60, 1);
+profiler.add(Profiler, "copyToClipboard").name("Copy result to clipboard");
+profiler.add(Profiler, "initProfiler").name("Initialize Profiler");
+profiler.add(Profiler, "state").name("Profiling state").disable(true).listen();
 
-profiler.add(timer, "profilerExecutionInSeconds", 1, 60, 1);
-profiler.add(timer, "startProfiler");
-profiler.add(timer, "stopProfiler");
-
-scene.init(canvas, gui);
-
-const fps = new Stats();
-
-fps.showPanel(0);
-
-document.body.appendChild(fps.dom);
+const fpsStats = new Stats();
+fpsStats.showPanel(0);
+changeScene();
+document.body.appendChild(fpsStats.dom);
 
 (function frame() {
-    fps.begin();
+    fpsStats.begin();
 
-    timer.lastTimestamp = timer.currentTimestamp;
-    timer.currentTimestamp = performance.now();
-
-    scene.update(timer.deltaTime / 1000);
-    scene.render();
-    timer.updateProfiler();
-
+    Timer.update();
+    if (sceneInitialized) {
+        scene.update(Timer.deltaTime / 1000);
+        scene.render();
+    }
+    Profiler.update(Timer.deltaTime);
     requestAnimationFrame(frame);
-    fps.end();
+
+    fpsStats.end();
     scrollFadeOut();
 })();
